Create README.md
The iOS app is like a thin client which fetches the data from the server and shows it on the screen as a list. 

It's a simple Master-Detail App and uses MapKit to load the map of the location. has caching facility with CoreData but it's not completely functional (radius matches need some work). 

The App pulls data from OpenMapquest API and uses reverseGeo coding to get location address.